pyproject.toml
[project]
name = "ai-chatbot"
version = "0.0.1"
description = "Config-driven local AI chatbot framework (modular, Ollama-first)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "PyYAML>=6.0.1",
    "requests>=2.31.0",
    "rich>=13.6.0",
    "httpx>=0.27.0",
    "pydantic>=2.7.3",
    "tenacity>=8.3.0",
    "colorama>=0.4.6"
]
